requirements.txt

Flask
Flask-SQLAlchemy
mysqlclient
pymysql
python-dotenv
groq
markdown2
bcrypt
pytz
